refactor(test): consolidate lifecycle/metrics/middleware-order tests into fastecho_test.go
All three were package fastecho server tests; merged into fastecho_test.go, no behavior change.
