--- lifecycle_test.go
@@ -1,30 +1,0 @@
-// Copyright © 2026 Ingka Holding B.V. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// You may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package fastecho
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestInitialize_HasProvidersAndShutdownIsSafe(t *testing.T) {
-	t.Setenv("OTEL_TRACES_EXPORTER", "none")
-	fe, err := Initialize(&Config{})
-	require.NoError(t, err)
-	require.NotNil(t, fe.server.Providers)
-	require.NoError(t, fe.Shutdown(context.Background()))
-}
